cell ternary, headercell util, expandedcontent styles
