+++ packages/table/src/Cell/HeaderCell/utils/index.ts
@@ -1,0 +1,1 @@
+export { getHeaderCellState } from './getHeaderCellState';
