+++ packages/table/src/ExpandedContent/ExpandedContent.styles.ts
@@ -1,8 +1,6 @@
-import { css, cx } from '@leafygreen-ui/emotion';
+import { css } from '@leafygreen-ui/emotion';
 import { Theme } from '@leafygreen-ui/lib';
 import { palette } from '@leafygreen-ui/palette';
-
-import { getCellContainerStyles } from '../Cell/Cell.styles';
 
 export const baseStyles = css`
   padding: 0;
@@ -16,6 +14,3 @@ export const expandedContentThemeStyles: Record<Theme, string> = {
     background-color: ${palette.gray.light3};
   `,
 };
-
-export const getContainerStyles = (theme: Theme) =>
-  cx(expandedContentThemeStyles[theme], getCellContainerStyles());
