+++ packages/table/src/ExpandedContent/ExpandedContent.tsx
@@ -8,7 +8,10 @@ import { LGIDS } from '../constants';
 import InternalRowBase from '../Row/InternalRowBase';
 import { useVirtualTableContext } from '../TableContext';
 
-import { baseStyles, getContainerStyles } from './ExpandedContent.styles';
+import {
+  baseStyles,
+  expandedContentThemeStyles,
+} from './ExpandedContent.styles';
 import { ExpandedContentProps } from './ExpandedContent.types';
 
 const ExpandedContent = <T extends RowData>({
@@ -39,7 +42,7 @@ const ExpandedContent = <T extends RowData>({
         className={cx(baseStyles)}
         data-lgid={LGIDS.cell}
       >
-        <div className={getContainerStyles(theme)}>
+        <div className={expandedContentThemeStyles[theme]}>
           <div>{content}</div>
         </div>
       </td>
